requirements.txt
# Core Data Processing
pandas==2.1.3
numpy==1.26.2

# Market Data APIs
alpaca-py==0.42.0
requests==2.31.0

# Data Storage and Performance
pyarrow==14.0.1

# Visualization and Analysis
matplotlib==3.8.2

# Development and Testing
pytest==7.4.3

# Optional: Jupyter for analysis notebooks
# jupyter==1.0.0
# notebook==7.0.6